Webpack badge handling
