+++ functions/azure/update-badges/index.js
@@ -1,0 +1,3 @@
+const { updateBadges } = require("../../../src/badges");
+
+exports.handler = context => updateBadges(context.done);
