+++ package.json
@@ -33,6 +33,7 @@
     "osm2obj": "^3.0.0",
     "pg": "^7.4.1",
     "pg-query-stream": "^1.1.1",
+    "pg.js": "^4.1.1",
     "require-env": "^0.2.1"
   },
   "devDependencies": {
