+++ src/badges/badges.js
@@ -1,0 +1,72 @@
+module.exports = {
+  pois: {
+    name: "On Point",
+    id: 3,
+    tiers: { 1: 500, 2: 2500, 3: 5000 }
+  },
+  buildings: {
+    name: "The Wright Stuff",
+    id: 4,
+    tiers: { 1: 100, 2: 500, 3: 1000 }
+  },
+  gpsTraces: {
+    name: "Field Mapper",
+    id: 5,
+    tiers: { 1: 10, 2: 50, 3: 100 }
+  },
+  roadKms: {
+    name: "On The Road Again",
+    id: 6,
+    tiers: { 1: 50, 2: 100, 3: 500 }
+  },
+  roadKmMods: {
+    name: "Long and Winding Road",
+    id: 7,
+    tiers: { 1: 50, 2: 100, 3: 500 }
+  },
+  waterways: {
+    name: "White Water Rafting",
+    id: 8,
+    tiers: { 1: 50, 2: 100, 3: 500 }
+  },
+  countries: {
+    name: "World Renown",
+    id: 9,
+    tiers: { 1: 5, 2: 10, 3: 25 }
+  },
+  tasks: {
+    name: "Task Champion",
+    id: 10,
+    tiers: { 1: 10, 2: 25, 3: 100 }
+  },
+  taskValidations: {
+    name: "Scrutinizer",
+    id: 11,
+    tiers: { 1: 25, 2: 100, 3: 250 }
+  },
+  josm: {
+    name: "Awesome JOSM",
+    id: 12,
+    tiers: { 1: 1, 2: 10, 3: 100 }
+  },
+  hashtags: {
+    name: "Mapathoner",
+    id: 13,
+    tiers: { 1: 5, 2: 20, 3: 50 }
+  },
+  daysInRow: {
+    name: "Consistency",
+    id: 14,
+    tiers: { 1: 5, 2: 20, 3: 50 }
+  },
+  daysTotal: {
+    name: "Year-Long Mapper",
+    id: 15,
+    tiers: { 1: 25, 2: 50, 3: 100 }
+  },
+  taskInvalidations: {
+    name: "High Standards",
+    id: 16,
+    tiers: { 1: 25, 2: 100, 3: 250 }
+  }
+};
