+++ src/badges/index.js
@@ -1,72 +1,150 @@
-module.exports = {
-  pois: {
-    name: "On Point",
-    id: 3,
-    tiers: { 1: 500, 2: 2500, 3: 5000 }
-  },
-  buildings: {
-    name: "The Wright Stuff",
-    id: 4,
-    tiers: { 1: 100, 2: 500, 3: 1000 }
-  },
-  gpsTraces: {
-    name: "Field Mapper",
-    id: 5,
-    tiers: { 1: 10, 2: 50, 3: 100 }
-  },
-  roadKms: {
-    name: "On The Road Again",
-    id: 6,
-    tiers: { 1: 50, 2: 100, 3: 500 }
-  },
-  roadKmMods: {
-    name: "Long and Winding Road",
-    id: 7,
-    tiers: { 1: 50, 2: 100, 3: 500 }
-  },
-  waterways: {
-    name: "White Water Rafting",
-    id: 8,
-    tiers: { 1: 50, 2: 100, 3: 500 }
-  },
-  countries: {
-    name: "World Renown",
-    id: 9,
-    tiers: { 1: 5, 2: 10, 3: 25 }
-  },
-  tasks: {
-    name: "Task Champion",
-    id: 10,
-    tiers: { 1: 10, 2: 25, 3: 100 }
-  },
-  taskValidations: {
-    name: "Scrutinizer",
-    id: 11,
-    tiers: { 1: 25, 2: 100, 3: 250 }
-  },
-  josm: {
-    name: "Awesome JOSM",
-    id: 12,
-    tiers: { 1: 1, 2: 10, 3: 100 }
-  },
-  hashtags: {
-    name: "Mapathoner",
-    id: 13,
-    tiers: { 1: 5, 2: 20, 3: 50 }
-  },
-  daysInRow: {
-    name: "Consistency",
-    id: 14,
-    tiers: { 1: 5, 2: 20, 3: 50 }
-  },
-  daysTotal: {
-    name: "Year-Long Mapper",
-    id: 15,
-    tiers: { 1: 25, 2: 50, 3: 100 }
-  },
-  taskInvalidations: {
-    name: "High Standards",
-    id: 16,
-    tiers: { 1: 25, 2: 100, 3: 250 }
+const { Writable } = require("stream");
+
+const async = require("async");
+const env = require("require-env");
+const { Pool } = require("pg");
+const QueryStream = require("pg-query-stream");
+
+const { NOOP } = require("..");
+const getBadges = require("./sum_check");
+const getDateBasedBadges = require("./date_check_total");
+const getSequentialBadges = require("./date_check_sequential");
+
+const pool = new Pool({
+  connectionString: env.require("DATABASE_URL")
+});
+
+pool.on("error", err => {
+  console.error("Unexpected error on idle client", err);
+  process.exit(-1);
+});
+
+class AbstractBadgeProcessor extends Writable {
+  constructor(pool) {
+    super({
+      objectMode: true
+    });
+
+    this.pool = pool;
+  }
+
+  updateBadges(userId, badges, callback) {
+    return async.each(
+      Object.keys(badges).map(x => badges[x]),
+      (badge, next) => {
+        this.pool.connect((err, client, release) => {
+          if (err) {
+            console.warn(err);
+            release();
+            return callback(err);
+          }
+
+          client.query(
+            `
+INSERT INTO badges_users (user_id, badge_id) VALUES (
+  $1, (SELECT id FROM badges WHERE category=$2 AND level=$3)
+)
+ON CONFLICT DO NOTHING
+            `,
+            [userId, badge.category, badge.level],
+            next
+          );
+
+          release();
+        });
+      },
+      err => {
+        if (err) {
+          console.warn(err);
+        }
+        return callback();
+      }
+    );
   }
 }
+
+class BadgeProcessor extends AbstractBadgeProcessor {
+  _write(row, _, callback) {
+    // TODO TM badges
+    const badges = getBadges({
+      buildings: Number(row.buildings_added),
+      pois: Number(row.pois_added),
+      roadKms: row.road_km_added,
+      roadKmMods: row.road_km_modified,
+      waterways: row.waterway_km_added,
+      josm: Number(row.josm_edits)
+    });
+
+    return this.updateBadges(row.user_id, badges, callback);
+  }
+}
+
+class DateBasedBadgeProcessor extends AbstractBadgeProcessor {
+  _write(row, _, callback) {
+    const { user_id: userId, timestamps } = row;
+
+    if (userId == null) {
+      return callback();
+    }
+
+    const badges = Object.assign(
+      {},
+      getDateBasedBadges(timestamps),
+      getSequentialBadges(timestamps)
+    );
+
+    return this.updateBadges(userId, badges, callback);
+  }
+}
+
+module.exports.updateBadges = callback => {
+  callback = callback || NOOP;
+
+  return async.parallel(
+    [
+      next =>
+        pool.connect((err, client, release) => {
+          if (err) {
+            throw err;
+          }
+
+          const done = () => {
+            release();
+
+            return next();
+          };
+
+          // TODO fetch users updated since the last time badges were updated
+          const query = new QueryStream("SELECT * FROM user_stats");
+
+          client.query(query).pipe(new BadgeProcessor(pool).on("finish", done));
+        }),
+      next =>
+        pool.connect((err, client, release) => {
+          if (err) {
+            throw err;
+          }
+
+          const done = () => {
+            release();
+
+            return next();
+          };
+
+          // TODO fetch data for users w/ changesets updated since the last time badges were updated
+          const query = new QueryStream(`
+  SELECT
+    user_id,
+    array_agg(DISTINCT date_trunc('day', created_at)) timestamps
+  FROM raw_changesets
+  GROUP BY user_id
+    `);
+
+          client
+            .query(query)
+            .pipe(new DateBasedBadgeProcessor(pool).on("finish", done));
+        })
+    ],
+    callback
+  );
+};
